+++ src/SqlBuilder/QueryPart/Column/Column.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace SqlBuilder\QueryPart;
+namespace SqlBuilder\QueryPart\Column;
 
 /**
  * Column
