+++ src/SqlBuilder/QueryPart/Column/IColumn.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace SqlBuilder\QueryPart;
+namespace SqlBuilder\QueryPart\Column;
 
 interface IColumn
 {
