+++ src/SqlBuilder/QueryPart/Condition/ICondition.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace SqlBuilder\QueryPart;
+namespace SqlBuilder\QueryPart\Condition;
 
 /**
  * SQL condition
