+++ src/SqlBuilder/QueryPart/Join/CrossJoinStmt.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace SqlBuilder\QueryPart;
+namespace SqlBuilder\QueryPart\Join;
 
 /**
  * Cross join
