+++ src/SqlBuilder/QueryPart/Join/IJoin.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace SqlBuilder\QueryPart;
+namespace SqlBuilder\QueryPart\Join;
 
 /**
  * SQL join
