+++ src/SqlBuilder/QueryPart/Join/JoinStmt.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace SqlBuilder\QueryPart;
+namespace SqlBuilder\QueryPart\Join;
 
 use SqlBuilder\Query;
