+++ src/SqlBuilder/QueryPart/Order/OrderStmt.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace SqlBuilder\QueryPart;
+namespace SqlBuilder\QueryPart\Order;
 
 /**
  * Order statement
